requirements.txt
# Dependências para o projeto Selenium Web Scraper

selenium==4.12.0
fpdf==1.7.2
pandas==2.0.3
matplotlib==3.7.3
python-dotenv==1.0.0